+++ fabrix/api.py
@@ -4,14 +4,18 @@ from fabrix.editor import insert_line, delete_line, prepend_line, append_line, r
 from fabrix.ioutil import read_file, read_local_file, write_file, write_local_file, copy_file, rsync, chown, chmod
 from fabrix.ioutil import remove_file, remove_directory, create_file, create_directory, name, warn, run, debug_print
 from fabrix.ioutil import is_file_exists, is_directory_exists, is_file_not_exists, is_directory_not_exists
+from fabrix.passwd import is_user_exists, is_user_not_exists, create_user, remove_user
+from fabrix.passwd import is_group_exists, is_group_not_exists, create_group, remove_group
+from fabrix.passwd import is_user_in_group, is_user_not_in_group, add_user_to_group, delete_user_from_group
+from fabrix.passwd import get_user_home_directory, add_user_ssh_authorized_keys
+from fabrix.render import render, render_template
+from fabrix.rpmyum import yum_install, yum_remove, yum_update
 from fabrix.system import systemctl_enable, systemctl_disable, systemctl_mask, systemctl_unmask
 from fabrix.system import systemctl_start, systemctl_stop, systemctl_reload, systemctl_restart
 from fabrix.system import systemctl_edit, is_reboot_required, reboot_and_wait, disable_selinux
 from fabrix.system import systemctl_get_default, systemctl_set_default, systemctl_preset
 from fabrix.system import localectl_set_locale, timedatectl_set_timezone
 from fabrix.system import get_virtualization_type
-from fabrix.rpmyum import yum_install, yum_remove, yum_update
-from fabrix.render import render, render_template
 
 # flake8: noqa